settings_org: Extract settings_components module.
