+++ tools/test-js-with-node
@@ -188,6 +188,7 @@ EXEMPT_FILES = make_set(
         "web/src/settings.js",
         "web/src/settings_account.js",
         "web/src/settings_bots.js",
+        "web/src/settings_components.js",
         "web/src/settings_display.js",
         "web/src/settings_emoji.ts",
         "web/src/settings_exports.ts",
