+++ web/src/settings_realm_user_settings_defaults.js
@@ -4,6 +4,7 @@ import * as audible_notifications from "./audible_notifications";
 import * as overlays from "./overlays";
 import {page_params} from "./page_params";
 import {realm_user_settings_defaults} from "./realm_user_settings_defaults";
+import * as settings_components from "./settings_components";
 import * as settings_display from "./settings_display";
 import * as settings_notifications from "./settings_notifications";
 import * as settings_org from "./settings_org";
@@ -49,7 +50,7 @@ export function update_page(property) {
     }
 
     const $input_elem = $container.find(`[name=${CSS.escape(property)}]`);
-    settings_org.set_input_element_value($input_elem, value);
+    settings_components.set_input_element_value($input_elem, value);
 }
 
 export function set_up() {
